+++ CollectDaprStats/GetNuGetPackageData.cs
@@ -58,6 +58,6 @@ namespace DaprStats
         public string PackageVersion { get; set; }
         public long? Downloads { get; set; }
         public DateTime CollectionDate { get ; set; }
-        public decimal CollectedOverNumberOfDays { get; set; }
+        public int CollectedOverNumberOfDays { get; set; }
     }
 }
