+++ CollectDaprStats/GetPythonPackageData.cs
@@ -1,6 +1,5 @@
 using Dapr.Workflow;
 using System.Net;
-using System.Net.Http.Json;
 
 namespace DaprStats
 {
@@ -60,7 +59,6 @@ namespace DaprStats
         public string PackageVersion { get; set; }
         public long? Downloads { get; set; }
         public DateTime CollectionDate { get; set; }
-        public decimal CollectedOverNumberOfDays { get; set; }
+        public int CollectedOverNumberOfDays { get; set; }
     }
-
 }
